requirements.txt
Flask==3.0.0
gunicorn==21.2.0
openai>=1.0.0
pdfplumber
sentence-transformers
faiss-cpu